Backend/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.4.2
boto==2.49.0
boto3==1.16.31
botocore==1.19.31
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-countries==7.2.1
django-dotenv==1.4.2
django-extra-fields==3.0.2
django-filter==2.4.0
django-fontawesome-5==1.0.18
django-heroku==0.3.1
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0
django-storages==1.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
FormEncode==2.0.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==3.1.1
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.8
Pillow==8.0.1
psycopg2==2.9.3
#psycopg2-binary==2.8.6
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
reportlab==3.6.2
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
# typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
zipp==3.4.0
